Netlist: fix visiting of variable declarations
